requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests>=2.31.0
python-dotenv==1.0.0
feedparser>=6.0.11
trafilatura>=1.9.0
tavily-python>=0.3.3
gunicorn==21.2.0
